Manager/requirements.txt
Flask
requests
gitpython
radon